split meta.hpp, add detected_t (used to define concepts)
