+++ include/nlohmann/detail/meta/void_t.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace nlohmann
+{
+namespace detail
+{
+template <typename...>
+using void_t = void;
+}
+}
